abstract config to class, update README, fix connectionm sensible defaults
