+++ example.env
@@ -11,3 +11,6 @@ WTFFMPEG_BEARER_TOKEN=
 
 # URL for a local LLM API like Ollama
 WTFFMPEG_LLM_API_URL=
+
+# Alternate directory for default --profile-dir
+WTFFMPEG_PROFILE_DIR=
