+++ pyproject.toml
@@ -46,4 +46,4 @@ where = ["src"]
 [options]
 include-package-data = true
 [tool.setuptools.package-data]
-"wtffmpeg" = ["profiles/*"]
+"wtffmpeg" = ["src/profiles/*"]
